src/requirements.txt
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
mplfinance==0.12.7a17
mysql-connector-python==8.0.25
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.3
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5
websocket-client==1.1.0